+++ pkg/sentry/fsimpl/gofer/lisafs_dentry.go
@@ -119,7 +119,7 @@ func (fs *filesystem) newLisafsDentry(ctx context.Context, ino *lisafs.Inode) (*
 			mode:      atomicbitops.FromUint32(uint32(ino.Stat.Mode)),
 			uid:       atomicbitops.FromUint32(uint32(fs.opts.dfltuid)),
 			gid:       atomicbitops.FromUint32(uint32(fs.opts.dfltgid)),
-			blockSize: atomicbitops.FromUint32(hostarch.PageSize),
+			blockSize: atomicbitops.FromUint32(uint32(hostarch.PageSize)),
 			readFD:    atomicbitops.FromInt32(-1),
 			writeFD:   atomicbitops.FromInt32(-1),
 			mmapFD:    atomicbitops.FromInt32(-1),
