+++ pkg/sentry/fsimpl/pipefs/pipefs.go
@@ -110,7 +110,7 @@ type inode struct {
 func newInode(ctx context.Context, fs *filesystem) *inode {
 	creds := auth.CredentialsFromContext(ctx)
 	return &inode{
-		pipe:  pipe.NewVFSPipe(false /* isNamed */, pipe.DefaultPipeSize),
+		pipe:  pipe.NewVFSPipe(false /* isNamed */, int64(pipe.DefaultPipeSize)),
 		ino:   fs.Filesystem.NextIno(),
 		uid:   creds.EffectiveKUID,
 		gid:   creds.EffectiveKGID,
@@ -153,7 +153,7 @@ func (i *inode) Stat(_ context.Context, vfsfs *vfs.Filesystem, opts vfs.StatOpti
 	defer i.attrMu.Unlock()
 	return linux.Statx{
 		Mask:     linux.STATX_TYPE | linux.STATX_MODE | linux.STATX_NLINK | linux.STATX_UID | linux.STATX_GID | linux.STATX_ATIME | linux.STATX_MTIME | linux.STATX_CTIME | linux.STATX_INO | linux.STATX_SIZE | linux.STATX_BLOCKS,
-		Blksize:  hostarch.PageSize,
+		Blksize:  uint32(hostarch.PageSize),
 		Nlink:    1,
 		UID:      uint32(i.uid),
 		GID:      uint32(i.gid),
