+++ pkg/sentry/fsimpl/tmpfs/named_pipe.go
@@ -32,7 +32,7 @@ type namedPipe struct {
 //   - fs.mu must be locked.
 //   - rp.Mount().CheckBeginWrite() has been called successfully.
 func (fs *filesystem) newNamedPipe(kuid auth.KUID, kgid auth.KGID, mode linux.FileMode, parentDir *directory) *inode {
-	file := &namedPipe{pipe: pipe.NewVFSPipe(true /* isNamed */, pipe.DefaultPipeSize)}
+	file := &namedPipe{pipe: pipe.NewVFSPipe(true /* isNamed */, int64(pipe.DefaultPipeSize))}
 	file.inode.init(file, fs, kuid, kgid, linux.S_IFIFO|mode, parentDir)
 	file.inode.nlink = atomicbitops.FromUint32(1) // Only the parent has a link.
 	return &file.inode
